Move v2 feature flag providers into their own namespace
